Backend

Wavefunction.jl

Wavefunction.jl defines the Wfn type and constructors. These constructors take in a Psi4 wavefunction object (PyObject) and convert necessary information to Julia data structures.

JuES.Wavefunction — Module

Module for storing and handling reference wavefunctions.

structs

Wfn -> holds info for disk based and in core computations.
DirectWfn -> holds info for integral direct computations.

JuES.Wavefunction.Wfn — Method

Wfn{T}(wfn::PyObject; unrestricted::Bool=false, diskbased::Bool=false, name::String = "default", df::Bool=false) where T

JuES.Wavefunction.Wfn — Type

Wfn

Data structure for storing integrals, MO coefficients, and misc information about a reference (HF) wavefunction.

Fields

nalpha::Int number of alpha electrons

nbeta::Int number of beta electrons

nvira::Int number of virtual functions of alpha spin

nvirb::Int number of virtual functions of beta spin

nmo::Int number of molecular orbitals

unrestricted::Bool whether or not the alpha and beta spatial extents are required to be the same.

Ca::Array{T,2} AO->MO coefficients for alpha MO's

Cb::Array{T,2} AO->MO coefficients for beta MO's

hao::Array{T,2} AO basis core hamiltonian (kinetic + potential)

epsa::Array{T,1} orbital eigenvalues for alpha MO's

epsb::Array{T,1} orbtial eigenvalues for beta MO's

ao_eri::Union{Array{T,4},DiskFourTensor} AO basis TEI

pqrs::Union{Array{T,4},DiskFourTensor} MO basis TEI (spin case AAAA)

pQrS::Union{Array{T,4},DiskFourTensor} MO basis TEI (spin case ABAB)

pQRs::Union{Array{T,4},DiskFourTensor} MO basis TEI (spin case ABBA)

PQRS::Union{Array{T,4},DiskFourTensor} MO basis TEI (spin case BBBB)

PqRs::Union{Array{T,4},DiskFourTensor} MO basis TEI (spin case BABA)

PqrS::Union{Array{T,4},DiskFourTensor} MO basis TEI (spin case BAAB)

Transformation.jl

Missing docstring.

Missing docstring for JuES.Transformation. Check Documenter's build log for details.

IntegralTransformation.jl

This module contain functions used to convert AO integrals into MO integrals. It can be used to get ERI arrays and Fock matrices.

JuES.IntegralTransformation — Module

JuES.IntegralTransformation

Module to handle integral transformations from AO to MO.

Functions:

get_eri   From a Wavefunction object, return a specified ERI array.
get_fock  From a Wavefunction object, return the Fock matrix.

JuES.IntegralTransformation.get_eri — Function

JuES.IntegralTransformation.get_eri(wfn::Wfn, eri_string::String, notation::String = "phys")

From a Wavefunction object, return a specified ERI array.

Arguments

wfn         Wavefunction object.
eri_string  String with length 4 identifying the type of ERI. 
            Characters must be (o, O, v, V). Each indicating 
            Occupied and Virtual for ALPHA and beta.

Kwargs

notation    {"chem", "phys"}
            Return the array in Chemist's or Physicists' notation.
            Default: "phys".

JuES.IntegralTransformation.get_fock — Function

JuES.IntegralTransformation.get_fock(wfn::Wfn, spin::String = "alpha")

From a Wavefunction object, return the Fock matrix.

Arguments

wfn  Wavefunction object.

Kwargs

spin    {"alpha", "a", "up", "beta", "b", "down"}
        String indicating the spin of the Fock matrix. 
        Case insensitive.

JuES.Transformation.tei_transform — Function

tei_transform(
              gao::Union{Array{Float64,4},Array{Float32,4}},
              C1::Union{Array{Float64,2},Array{Float32,2}},
              C2::Union{Array{Float64,2},Array{Float32,2}},
              C3::Union{Array{Float64,2},Array{Float32,2}},
              C4::Union{Array{Float64,2},Array{Float32,2}},
              name::String
              )

Transform gao to the MO basis defined by C1,C2,C3,C4.

DF.jl

JuES.DF — Module

Module for Density Fitting (DF) utility routines

exports: functions setup_df