Coupled Cluster

JuES.CoupledCluster — Module

Module for running CC computations in Julia.

Implemented –> RCCD, RCCSD, DF-RCCD

CCD

JuES.CoupledCluster.RCCD — Module

JuES.CoupledCluster.RCCD

performs coupled cluster doubles (CCD) computations.

Functions

JuES.CoupledCluster.RCCD.do_rccd

JuES.CoupledCluster.RCCD.do_rccd — Function

do_rccd

Applies RCCD equations to the input Wfn object. Disk based versus in-core algorithm is selected based on the type of atomic orbitals in Wfn.uvsr. –-

paramters

refWfn::Wfn -> Wfn object to which the RCCD equations will be applied.

maxit::Int -> maximum number of coupled cluster iterations.

doprint::Bool=false -> whether or not to print energy and timing information to stdout.

output

ccenergy::Float -> final RCCD energy.

CCSD

JuES.CoupledCluster.RCCSD — Module

JuES.CoupledCluster.RCCSD

Functions

JuES.CoupledCluster.RCCSD.do_rccsd

JuES.CoupledCluster.RCCSD.do_rccsd — Function

JuES.CoupledCluster.RCCSD.do_rccsd(refWfn::Wfn; kwargs...)

Arguments

refWfn::Wfn

reference wavefunction to compute RCCSD on.

kwargs::Any

keyword arguments for various options.

Kwargs

:doprint::Bool  print or no? deprecated
:maxit::Int     max number of iterations for RCCSD equations
:return_T::Bool do return the converged cluster amplitudes?
:diis::Bool     do DIIS extrapolation? currently dummy